move file into games folder
