+++ src/commands/index.ts
@@ -1,6 +1,6 @@
 import { Cat } from "@commands/fun/Cat";
 import { Cheese } from "@commands/fun/Cheese";
-import { Chess } from "@commands/fun/Chess";
+import { Chess } from "@src/commands/games/Chess";
 import { Coinflip } from "@commands/fun/Coinflip";
 import { Fact } from "@commands/fun/Fact";
 import { Ferret } from "@commands/fun/Ferret";
